requirements.txt
Flask==3.0.3
google-generativeai==0.5.2
grpcio-status==1.62.2
numpy==1.26.4
pillow==10.3.0
pipreqs==0.5.0
python-dotenv==1.0.1
reflex==0.4.8.post1
spotipy==2.23.0
